requirements.txt
numpy
Pillow
tensorflow==2.3.0
tensorflow-gpu==2.3.0
albumentations==0.5.2
matplotlib
opencv-python
scipy
scikit-learn

